==> code/3_normoxia.Rmd <==
---
title: "EIF3E and EIF3D-mediated translational regulation in normoxia"
author: "`r Sys.Date()`"
date: ""
output: html_document
editor_options: 
  chunk_output_type: console
---

# Figure: Overlap between functions of eIF3e and eIF3d in normoxia

```{r setup, echo=FALSE}
knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning=FALSE)
```

```{r install packages, echo=FALSE}

library(here)
library(janitor)
library(tximport)
library(tidyverse)
library(EnhancedVolcano)
library(msigdbr)
library(viridis)
library(gridExtra)
library(ggthemes)
library(ggrepel)
library(scales)
library(DESeq2)
library(limma)
library(clusterProfiler)
library(plotly)
library(ggokabeito)
library(ggvenn)
library(ggpubr)
library(DT)
library(edgeR)
library(ggprism)
library(ggExtra)
library(reshape2)
library(pheatmap)
library(enrichplot)
library(GeneOverlap)
library(viridis)


```

```{r}

theme_set(theme_prism())


a209_col = "#CC79A7"
si3e_col = "#56B4E9"
si3d_col = "#E69F00"
```

## load metadata
```{r}
metadata <- read_csv(here("counts/calculate_te/metadata.csv"))
```

## filter metadata for samples of interest
```{r}
metadata_norm <- metadata %>% 
  filter(treatment %in% c("sictrl","si3d","si3e")) %>%
  filter(condition == "normoxia")
```

## add gene id to symbol key

```{r}
geneInfo <- read_csv(here("accessories","gencode.v26.primary.info.csv.gz"), col_names = F) %>%
  filter(!str_detect(X2, "pre_"))

colnames(geneInfo) <- c("gene_id","transcript_id","biotype","symbol")

#map transcripts to gene
tx2gene <- geneInfo[,c(2,1)]
gene_symbol <- geneInfo %>% dplyr::select(gene_id, symbol) %>% unique()

```

## Separate ribo counts, rna counts, and te counts
```{r}
raw_counts <- read_csv(here("counts/calculate_te/raw_filtered_allCounts.csv")) %>% 
  column_to_rownames("gene_id")

raw_counts_edit <- raw_counts %>%
  dplyr::select(!matches("DMSO|a209"))

ribo_counts <- as.matrix(raw_counts_edit[,grep(pattern = "ribo",colnames(raw_counts_edit))])
rna_counts <- as.matrix(raw_counts_edit[,grep(pattern = "rna",colnames(raw_counts_edit))])


rna_ribo_counts <- as.matrix(raw_counts_edit)

```

## Delta TE eIF3d KD vs. sicontrol

```{r}

#organize metadata
metadata_norm_3d <- metadata_norm %>% 
  filter(treatment== "sictrl" | treatment=="si3d") %>% 
  filter(condition=="normoxia") %>% 
  mutate("name" = paste(condition, treatment, rep, library, sep = "_"))

metadata_norm_3d$library <- relevel(factor(metadata_norm_3d$library), ref = "rna")
metadata_norm_3d$treatment <- relevel(factor(metadata_norm_3d$treatment), ref = "sictrl")

levels(metadata_norm_3d$library)
levels(metadata_norm_3d$treatment)

names <- as.character(metadata_norm_3d$name)

comparison_3d <- rna_ribo_counts[, colnames(rna_ribo_counts) %in% names]


ddsMat_3d <- DESeqDataSetFromMatrix(
  countData = comparison_3d,
  colData = metadata_norm_3d,
  design = ~treatment + library + treatment:library
  )

ddsMat_3d <- DESeq(ddsMat_3d)
resultsNames(ddsMat_3d)
res_te_3d <- results(ddsMat_3d, name=resultsNames(ddsMat_3d)[4], cooksCutoff=F, independentFiltering=F)


res_te_3d <- lfcShrink(dds = ddsMat_3d, coef = resultsNames(ddsMat_3d)[4], res = res_te_3d) %>% 
  as.data.frame %>% 
  rownames_to_column(var = "gene_id") %>%
  inner_join(gene_symbol, .)
  # filter(!symbol %in% nonpolyA)  - check if this is necessary, should already be filtered

write_csv(res_te_3d, here("counts/normoxia_sictrl_3d.csv"))
```

## Delta TE eIF3e KD vs. sicontrol

```{r}

#organize metadata
metadata_norm_3e <- metadata_norm %>% 
  filter(treatment== "sictrl" | treatment=="si3e") %>% 
  filter(condition=="normoxia") %>% 
  mutate("name" = paste(condition, treatment, rep, library, sep = "_"))

metadata_norm_3e$library <- relevel(factor(metadata_norm_3e$library), ref = "rna")
metadata_norm_3e$treatment <- relevel(factor(metadata_norm_3e$treatment), ref = "sictrl")

levels(metadata_norm_3e$library)
levels(metadata_norm_3e$treatment)

names <- as.character(metadata_norm_3e$name)

comparison_3e <- rna_ribo_counts[, colnames(rna_ribo_counts) %in% names]


ddsMat_3e <- DESeqDataSetFromMatrix(
  countData = comparison_3e,
  colData = metadata_norm_3e,
  design = ~treatment + library + treatment:library
  )

ddsMat_3e <- DESeq(ddsMat_3e)
resultsNames(ddsMat_3e)
res_te_3e <- results(ddsMat_3e, name=resultsNames(ddsMat_3e)[4], cooksCutoff=F, independentFiltering=F)


res_te_3e <- lfcShrink(dds = ddsMat_3e, coef = resultsNames(ddsMat_3e)[4], res = res_te_3e) %>% 
  as.data.frame %>% 
  rownames_to_column(var = "gene_id") %>%
  inner_join(gene_symbol, .)
  # filter(!symbol %in% nonpolyA) 


write_csv(res_te_3e, here("counts/normoxia_sictrl_3e.csv"))

```

## Import TE 
Calculated previously in calculate_te.Rmd
allgene TE has normalized counts for RNA and RIBO as well
```{r}
all_gene_te <- read_csv(here("counts/calculate_te/complete_te_list.csv"))

all_gene_te_long <- all_gene_te %>% 
  melt() %>% 
  separate(variable, into = c("condition", "treatment", "rep", "type"))

slimmed_te <- all_gene_te_long %>% 
  filter(condition=="normoxia") %>% 
  filter(treatment != "a209") %>% 
  filter(treatment != "DMSO")

```

## Visualize TE for example genes and how they change with 3e and 3d KD

```{r}

d_genes <- c("CLDN9", "CLIC1", "PRDM4")
e_genes <- c("APLP2", "PABPC1", "TOR1AIP2")
common <- c("ATF4", "EIF4EBP2", "GSK3B", "MED13L", "PCBP2", "XBP1")

```

Plot genes - common
```{r}

slimmed_te$treatment <- relevel(factor(slimmed_te$treatment), ref = "sictrl")

common_te <- slimmed_te %>% 
  filter(symbol %in% common) %>% 
  filter(type=="TE")


common_plot <-
ggstripchart(data = common_te, x = "treatment", y = "value", color = "treatment", facet.by = "symbol", scales="free_y", palette = c("black",si3d_col,si3e_col), position = position_jitterdodge(1.2)) + 
  ylab(expression("log"[2]*" TE (ribo/rna)"))
  
common_plot

ggsave(width = 5.5, height = 3.75, filename = here("plots/commongenes.pdf"))


```

Plot genes - d only
```{r}
d_te <- slimmed_te %>% 
  filter(symbol %in% d_genes) %>% 
  filter(type=="TE")


d_plot <-
ggstripchart(data = d_te, x = "treatment", y = "value", color = "treatment", facet.by = "symbol", scales="free_y", palette = c("black",si3d_col,si3e_col), position = position_jitterdodge(1.2)) + 
  ylab(expression("log"[2]*" TE (ribo/rna)"))
  
d_plot

ggsave(width = 5.5, height = 2.5, filename = here("plots/dgenes.pdf"))

```


Plot genes - e only

```{r}
e_te <- slimmed_te %>% 
  filter(symbol %in% e_genes) %>% 
  filter(type=="TE")


e_plot <-
ggstripchart(data = e_te, x = "treatment", y = "value", color = "treatment", facet.by = "symbol", scales="free_y", palette = c("black",si3d_col,si3e_col), position = position_jitterdodge(1.2)) + 
  ylab(expression("log"[2]*" TE (ribo/rna)"))
  
e_plot

ggsave(width = 5.5, height = 2.5, filename = here("plots/egenes.pdf"))

```

heatmap of genes

```{r}

genes_of_interest<-
slimmed_te %>% 
  filter(symbol %in% e_genes|symbol %in% d_genes|symbol %in% common) %>% 
  filter(type=="TE") %>% 
  dplyr::select(symbol, treatment, rep, value) %>%
  pivot_wider(names_from = c(treatment,rep) , values_from = value) %>%
  column_to_rownames("symbol")

mynames <- colnames(genes_of_interest)[c(7:9,4:6,1:3)]

mygeneorder <- c("ATF4", "EIF4EBP2", "GSK3B", "MED13L", "PCBP2", "XBP1", "APLP2", "PABPC1", "TOR1AIP2","CLDN9", "PRDM4","CLIC1")

d <- genes_of_interest[mygeneorder,]

my_col <- data.frame(sample = rep(c("siCtrl", "si3e","si3d"), each = 3))
row.names(my_col) <- colnames(d)
levels(factor(my_col$sample))
my_col$sample <- factor(factor(my_col$sample), levels = c("siCtrl", "si3e","si3d"))

my_row <- data.frame(gene = c(rep("both",6),rep("3e-specific",3),rep("3d-specific",3)))
row.names(my_row) <- rownames(d)        
my_row$gene <- factor(factor(my_row$gene), levels = c("both", "3e-specific","3d-specific"))

paletteLength <- 50
myColor <- colorRampPalette(c("darkblue", "white", "darkred"))(paletteLength)
myBreaks <- c(seq(min(d), 0, length.out=ceiling(paletteLength/2) + 1), 
              seq(max(d)/paletteLength, max(d), length.out=floor(paletteLength/2)))

example_genes_heatmap<- 
pheatmap(mat = d,
         scale = "row",
         cluster_cols = F, 
         cluster_rows = F,
         annotation_col = my_col,
         annotation_row = my_row,
         color = myColor, fontsize = 10
          )

example_genes_heatmap

ggsave(here("plots/example_genes.pdf"), width = 4, height = 4)
```



Visualize e and d specifically: controls

```{r}
e_and_d <-slimmed_te %>% 
  filter(symbol=="EIF3E" | symbol=="EIF3D")

ggstripchart(data = e_and_d, x = "treatment", y = "value", color = "treatment", scales="free_y", palette = c("black",si3d_col,si3e_col), position = position_jitterdodge(1.2))+
  facet_wrap(~symbol+type)


#Replot as averages with standard deviations

bars_e_d_norm <- e_and_d %>% 
  filter(type != "TE")

bars_e_d_summary <- bars_e_d_norm %>%
  group_by(symbol, type, treatment) %>%
  summarise(
    mean_value = mean(value),
    sd = sd(value)  # Standard deviation
  )

bars_e_d_summary$type <- factor(bars_e_d_summary$type, levels=c("rna","ribo"))

ggplot(bars_e_d_summary, aes(x = treatment, y = mean_value, fill = treatment)) +
  geom_bar(stat = "identity", position = position_dodge()) +
  geom_errorbar(aes(ymin = mean_value - sd, ymax = mean_value + sd), 
                width = 0.2, position = position_dodge(0.9)) +
  facet_wrap(~ symbol + type) +
  scale_fill_manual(values = c("grey",si3d_col,si3e_col)) +
  # theme_cowplot()+
    theme(legend.position = "none")

ggsave(here("plots/e_d_rna_ribo.pdf"), width = 4, height = 4.5)

```


## Venn diagram of 3e/d gene overlaps in normoxia
```{r compare genes}
te_3d_up <- res_te_3d %>% 
  filter(padj < 0.05 & log2FoldChange > 0) %>% pull(gene_id)

te_3d_down <- res_te_3d %>% 
  filter(padj < 0.05 & log2FoldChange < 0) %>% pull(gene_id)

te_3e_up <- res_te_3e %>% 
  filter(padj < 0.05 & log2FoldChange > 0) %>% pull(gene_id)

te_3e_down <- res_te_3e %>% 
  filter(padj < 0.05 & log2FoldChange < 0) %>% pull(gene_id)


norm_te_up <-
  list(
    te_3d_up = res_te_3d %>% 
  filter(padj < 0.05 & log2FoldChange > 0) %>% pull(gene_id),
  te_3e_up = res_te_3e %>% 
  filter(padj < 0.05 & log2FoldChange > 0) %>% pull(gene_id)
  )

norm_te_down <-
  list(
    te_3d_down = res_te_3d %>% 
  filter(padj < 0.05 & log2FoldChange < 0) %>% pull(gene_id),
  te_3e_down = res_te_3e %>% 
  filter(padj < 0.05 & log2FoldChange < 0) %>% pull(gene_id)
  )


ggvenn(norm_te_up, auto_scale = T, fill_color = c(si3d_col,si3e_col), text_size = 2)

ggsave(width = 2, height = 2, filename = here("plots/vennup.pdf"))


ggvenn(norm_te_down, auto_scale = T, fill_color = c(si3d_col,si3e_col), text_size = 2)

ggsave(width = 2, height = 2, filename = here("plots/venndown.pdf"))


```

### Test significance of overlap
Create a matrix that compares each condition for gene overlap analysis.

#universe number calculation

```{r}
all_normox_expressed <- rbind(
res_te_3e,
res_te_3d)

universe <- unique(all_normox_expressed$gene_id)


```


# down genes overlap
```{r}

norm_te_down_df <- norm_te_down %>% 
  purrr::imap_dfr(~ tibble(list_name = .y, gene_name = .x))

gom.allsamples_down <- split(norm_te_down_df$gene_name, norm_te_down_df$list_name)

#double check genome size here

gom.all_down <- newGOM(gom.allsamples_down, gom.allsamples_down, genome.size = 11635)

norm_all.odds.matrix_down <- getMatrix(gom.all_down, "odds")
norm_all.pval.matrix_down <- getMatrix(gom.all_down, "pval")

norm_all.pval.matrix_down
```

# up genes overlap
```{r}

norm_te_up_df <- norm_te_up %>% 
  purrr::imap_dfr(~ tibble(list_name = .y, gene_name = .x))

gom.allsamples_up <- split(norm_te_up_df$gene_name, norm_te_up_df$list_name)

#double check genome size here

gom.all_up <- newGOM(gom.allsamples_up, gom.allsamples_up, genome.size = 11635)

norm_all.odds.matrix_up <- getMatrix(gom.all_up, "odds")
norm_all.pval.matrix_up <- getMatrix(gom.all_up, "pval")

norm_all.pval.matrix_up
```


### Density scatter plot

```{r}
norm_sig_te <- data.frame(
  gene_id = res_te_3d$gene_id,
  symbol = res_te_3d$symbol,
  te_3d = res_te_3d$log2FoldChange,
  te_3e = res_te_3e$log2FoldChange
) %>%
  filter(gene_id %in% c(te_3e_up,
                        te_3e_down,
                        te_3d_up,
                        te_3d_down)) 

scatter <-
ggscatter(data = norm_sig_te,
          x = "te_3d",
          y = "te_3e",
          xlab= "si-eIF3d TE",
          ylab="si-eIF3e TE",
          alpha = 0.25,
          title = "TE changes with 3d KD vs 3e KD", 
          xlim=c(-2,2), ylim = c(-2,2),
  add = "reg.line")+
   stat_cor(
   aes(label =
         paste(
           after_stat(rr.label),
           ..p.label..,
           sep = "~`,`~")
       ) + theme_prism()
   )+
  stat_regline_equation(
    aes(label =  paste(..eq.label.., after_stat(rr.label), sep = "~~~~")),
    label.x = -1.5, label.y = 1.5
  ) +
  geom_vline(xintercept=0, linetype='dotted') +
  geom_hline(yintercept=0, linetype='dotted') +
  theme_prism() +
  guides(color = FALSE)

scatter

ggsave(width = 5, height = 5, filename = here("plots/normoxia_quadrants.pdf"))

```


## Pathway analysis setup

```{r prep GSEA analysis}

eif3d_te <- res_te_3d$log2FoldChange
names(eif3d_te) <- res_te_3d$symbol

eif3d_te <- sort(eif3d_te, decreasing = TRUE)

eif3d_te <- eif3d_te[!duplicated(names(eif3d_te))]

eif3e_te <- res_te_3e$log2FoldChange
names(eif3e_te) <- res_te_3e$symbol

eif3e_te <- sort(eif3e_te, decreasing = TRUE)

eif3e_te <- eif3e_te[!duplicated(names(eif3e_te))]


m_df_h <- msigdbr(species = "Homo sapiens") %>%
  filter(gs_cat == "H") %>% 
  dplyr::select(gs_name, gene_symbol)

m_df_c2 <- msigdbr(species = "Homo sapiens") %>%
  filter(gs_cat == "C2" & gs_subcat != "CGP") %>% 
  dplyr::select(gs_name, gene_symbol)

```

### Hallmark pathways
```{r GSEA Hallmarks}

eif3d_te_h <- GSEA(
  geneList = eif3d_te,
  eps = 0,
  pAdjustMethod = "fdr",
  pvalueCutoff = 1,
  minGSSize = 20,
  maxGSSize = 1000,
  TERM2GENE = m_df_h
  )

eif3e_te_h <- GSEA(
  geneList = eif3e_te,
  eps = 0,
  pAdjustMethod = "fdr",
  pvalueCutoff = 1,
  minGSSize = 20,
  maxGSSize = 1000,
  TERM2GENE = m_df_h
  )

norm_te_h <- dplyr::full_join(
  eif3d_te_h@result %>%
  dplyr::select(ID,NES,pvalue,qvalue) %>%
  rename_with(~ paste0("eif3d_", .), is.numeric),
  eif3e_te_h@result %>%
  dplyr::select(ID,NES,pvalue,qvalue) %>%
  rename_with(~ paste0("eif3e_", .), is.numeric),
  by = "ID") %>% 
  filter(eif3d_pvalue < 0.05 | eif3e_pvalue < 0.05)  %>%
  dplyr::arrange(-eif3e_NES) %>%
  mutate(delta = eif3e_NES - eif3d_NES)



norm_h_table <- norm_te_h %>% 
  mutate(direction = ifelse(eif3d_NES < 0 & eif3e_NES < 0, "down_both",
                            ifelse(eif3d_NES > 0 & eif3e_NES > 0, "up_both",
                            ifelse(eif3d_NES > 0 & eif3e_NES < 0, "up_3d_down_3e",
                            ifelse(eif3d_NES < 0 & eif3e_NES > 0, "up_3e_down_3d", "NA")))))

write_csv(norm_h_table, here("counts/norm_h_table.csv"))

```



# Determine hypoxia pathway enrichment
```{r}

#3e
gseaplot(eif3e_te_h, geneSetID= "HALLMARK_HYPOXIA", color.line = si3e_col, color.vline = si3e_col, ylim=c(-0.4,0.4))


ggsave(here("plots/eIF3e_hypoxiapathway.pdf"), width=6, height=5)


gseaplot(eif3d_te_h, geneSetID= "HALLMARK_HYPOXIA", color.line = si3d_col, color.vline = si3d_col)
ggsave(here("plots/eIF3d_hypoxiapathway.pdf"), width=6, height=5)


```


### KEGG pathways
```{r GSEA KEGG}
eif3d_te_kegg <- GSEA(
  geneList = eif3d_te,
  eps = 0,
  pAdjustMethod = "fdr",
  pvalueCutoff = 1,
  minGSSize = 20,
  maxGSSize = 1000,
  TERM2GENE = m_df_c2 %>%
  filter(grepl('^KEGG', gs_name))
  )

eif3e_te_kegg <- GSEA(
  geneList = eif3e_te,
  eps = 0,
  pAdjustMethod = "fdr",
  pvalueCutoff = 1,
  minGSSize = 20,
  maxGSSize = 1000,
  TERM2GENE = m_df_c2 %>%
  filter(grepl('^KEGG', gs_name))
  )

norm_te_kegg <- dplyr::full_join(
  eif3d_te_kegg@result %>%
  dplyr::select(ID,NES,pvalue,qvalue) %>%
  rename_with(~ paste0("eif3d_", .), is.numeric),
  eif3e_te_kegg@result %>%
  dplyr::select(ID,NES,pvalue,qvalue) %>%
  rename_with(~ paste0("eif3e_", .), is.numeric),
  by = "ID") %>% 
  filter(eif3d_qvalue < 0.05 | eif3e_qvalue < 0.05) %>%
  dplyr::arrange(-eif3e_NES) %>%
  mutate(delta = eif3e_NES - eif3d_NES)

norm_te_kegg$ID <- gsub(pattern = "KEGG_", replacement = "", x = norm_te_kegg$ID)

ggscatter(data = norm_te_kegg,
          x = "eif3d_NES",
          y = "eif3e_NES",
          label = "ID",
          repel = TRUE,
          title = "Kegg - Normoxia",
          font.label = 7) +
  scale_color_continuous(type = "viridis") +
  geom_vline(xintercept = 0, linetype = "dashed") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  ylim(-2.5,2.5) +
  xlim(-2.5,2.5) + theme_prism()


DT::datatable(norm_te_kegg) %>%
  formatRound(which(sapply(norm_te_kegg,is.numeric)), digits = 3) 


ggsave(width = 4, height = 4, filename = here("plots/normoxia_kegg.pdf"))

norm_kegg_table <- norm_te_kegg %>% 
  mutate(direction = ifelse(eif3d_NES < 0 & eif3e_NES < 0, "down_both",
                            ifelse(eif3d_NES > 0 & eif3e_NES > 0, "up_both",
                            ifelse(eif3d_NES > 0 & eif3e_NES < 0, "up_3d_down_3e",
                            ifelse(eif3d_NES < 0 & eif3e_NES > 0, "up_3e_down_3d", "NA")))))

write_csv(norm_kegg_table, here("counts/norm_kegg_table.csv"))


```
